+++ src/lanpartydb/reading.py
@@ -1,0 +1,80 @@
+"""
+lanpartydb.reading
+~~~~~~~~~~~~~~~~~~
+
+Data reading
+
+:Copyright: 2024 Jochen Kupperschmidt
+:License: MIT
+"""
+
+from decimal import Decimal
+from pathlib import Path
+import tomllib
+from typing import Any
+
+from .models import Links, Location, Party, Series, Website
+
+
+# series
+
+
+def read_series_list_from_toml_file(filename: Path) -> list[Series]:
+    """Read list of series from a TOML file."""
+    toml = filename.read_text()
+    return read_series_list_from_toml(toml)
+
+
+def read_series_list_from_toml(toml: str) -> list[Series]:
+    """Read list of series from a TOML document."""
+    data = _load_toml(toml)
+    return _read_series_list_from_dict(data)
+
+
+def _read_series_list_from_dict(data: dict[str, Any]) -> list[Series]:
+    """Read list of series from a dictionary."""
+    return [Series(**item) for item in data['series']]
+
+
+# party
+
+
+def read_party_from_toml_file(filename: Path) -> Party:
+    """Read party from a TOML file."""
+    toml = filename.read_text()
+    return read_party_from_toml(toml)
+
+
+def read_party_from_toml(toml: str) -> Party:
+    """Read party from a TOML document."""
+    data = _load_toml(toml)
+    return _read_party_from_dict(data)
+
+
+def _read_party_from_dict(party_dict: dict[str, Any]) -> Party:
+    """Read party from a dictionary."""
+    location_dict = party_dict.pop('location', None)
+    if location_dict:
+        party_dict['location'] = Location(**location_dict)
+
+    links_dict = party_dict.pop('links', None)
+    if links_dict:
+        website_obj = links_dict.pop('website', None)
+        if website_obj:
+            if isinstance(website_obj, dict):
+                website = Website(
+                    url=website_obj['url'],
+                    offline=website_obj.get('offline', False),
+                )
+            else:
+                website = Website(url=website_obj)
+            party_dict['links'] = Links(website=website)
+
+    return Party(**party_dict)
+
+
+# util
+
+
+def _load_toml(toml: str) -> str:
+    return tomllib.loads(toml, parse_float=Decimal)